requirements.txt
torch>=2.1.0
transformers>=4.40.0
peft>=0.10.0
datasets>=2.18.0
accelerate>=0.29.0
scikit-learn>=1.3.0
numpy>=1.24.0
pyyaml>=6.0
tqdm>=4.65.0
sentencepiece>=0.1.99
protobuf
evaluate>=0.4.0
scipy
